Added overload for easier use with other platforms. Renamed NinaRazorTemplateEngine to RazorTemplateEngine and TemplateEngine to ITemplateEngine.
