+++ Nina.ViewEngines.Razor/Nina.ViewEngines.Razor.csproj
@@ -47,7 +47,7 @@
   <ItemGroup>
     <Compile Include="RazorConfig.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="NinaRazorTemplateEngine.cs" />
+    <Compile Include="RazorTemplateEngine.cs" />
     <Compile Include="RazorTemplate.cs" />
   </ItemGroup>
   <ItemGroup>
